requirements.txt
python-can
coverage
